fix(code-cleanup): comments from review
