+++ config/mocks/wallet-options-v4.json
@@ -19,8 +19,6 @@
       "featureFlags": {
         "cEURRewards": true,
         "eddInterestFileUpload": true,
-        "legacyMobilePairing": false,
-        "legacyWalletRecovery": false,
         "pollForMagicLinkData": true,
         "recurringBuys": true,
         "signupCountry": true
