+++ packages/blockchain-wallet-v4-frontend/src/scenes/Login/VerifyMagicLink/index.tsx
@@ -1,13 +1,12 @@
 import React from 'react'
 import { connect } from 'react-redux'
-import { bindActionCreators } from 'redux'
 
-import { Text } from 'blockchain-info-components'
-import { actions, selectors } from 'data'
+import { selectors } from 'data'
 
 import { Props } from '..'
 import Error from './template.error'
 import Loading from './template.loading'
+import VerifyDevicePrompt from './template.prompt'
 import Success from './template.success'
 
 class VerifyMagicLink extends React.PureComponent<Props> {
@@ -15,8 +14,8 @@ class VerifyMagicLink extends React.PureComponent<Props> {
     return this.props.data.cata({
       Failure: (val) => <Error error={val} />,
       Loading: () => <Loading />,
-      NotAsked: () => <Success {...this.props} success={false} />,
-      Success: () => <Success {...this.props} success />
+      NotAsked: () => <VerifyDevicePrompt {...this.props} />,
+      Success: () => <Success />
     })
   }
 }
